Battery charge diagnostics
- Moves system_power_management functions into separate CU system_power.cpp
- Moves system_power_management_update() call to system thread
- Adds thread-safety to
  system_power_management_update()/system_power_management_init()
